styling

+++ server/routes/convo.router.js
@@ -4,32 +4,32 @@ const router = express.Router();
 
 // POST
 router.post('/', (req, res) => {
-    if (req.isAuthenticated()) {
-      const queryText = `INSERT INTO comments ("comment", "topic_id", "subtopic_id", "user_id")
+  if (req.isAuthenticated()) {
+    const queryText = `INSERT INTO comments ("comment", "topic_id", "subtopic_id", "user_id")
                       VALUES ($1, $2, $3, $4) RETURNING "comment";`;
-      pool.query(queryText, [req.body.comment, req.body.topic, req.body.subtopic, req.user.id])
-        .then(() => { res.sendStatus(201); })
-        .catch((err) => {
-          console.log('Error completing POST comment query', err);
-          res.sendStatus(500);
-        })
-    } else {
-      res.sendStatus(403);
-    }
-  });
+    pool.query(queryText, [req.body.comment, req.body.topic, req.body.subtopic, req.user.id])
+      .then(() => { res.sendStatus(201); })
+      .catch((err) => {
+        console.log('Error completing POST comment query', err);
+        res.sendStatus(500);
+      })
+  } else {
+    res.sendStatus(403);
+  }
+});
 
 // GET
 
 router.get('/', (req, res) => {
   if (req.isAuthenticated()) {
 
-  const queryText = 'SELECT * FROM comments';
-  pool.query(queryText)
-    .then((result) => { res.send(result.rows); })
-    .catch((err) => {
-      console.log('Error completing GET comments query', err);
-      res.sendStatus(500);
-    });
+    const queryText = 'SELECT * FROM comments';
+    pool.query(queryText)
+      .then((result) => { res.send(result.rows); })
+      .catch((err) => {
+        console.log('Error completing GET comments query', err);
+        res.sendStatus(500);
+      });
   } else {
     res.sendStatus(403);
   }
@@ -37,14 +37,13 @@ router.get('/', (req, res) => {
 
 router.get('/:id', (req, res) => {
   if (req.isAuthenticated()) {
-
-  const queryText = 'SELECT * FROM comments WHERE "subtopic_id"=$1';
-  pool.query(queryText, [req.params.id])
-    .then((result) => { res.send(result.rows); })
-    .catch((err) => {
-      console.log('Error completing GET comments query', err);
-      res.sendStatus(500);
-    });
+    const queryText = 'SELECT * FROM comments WHERE "subtopic_id"=$1;'
+    pool.query(queryText, [req.params.id])
+      .then((result) => { res.send(result.rows); })
+      .catch((err) => {
+        console.log('Error completing GET comments query', err);
+        res.sendStatus(500);
+      });
   } else {
     res.sendStatus(403);
   }
@@ -52,36 +51,42 @@ router.get('/:id', (req, res) => {
 
 // DELETE
 
-router.delete('/:id', (req,res) => {
-    console.log('delete from SQL')
-
+router.delete('/:id', (req, res) => {
+  if (req.isAuthenticated()) {
     const deleteInfo = req.params.id;
     pool.query('DELETE FROM "comments" WHERE "id"=$1;', [deleteInfo])
-        .then((result) => {
-            res.sendStatus(200);
-    }).catch((error) => {
+      .then((result) => {
+        res.sendStatus(200);
+      }).catch((error) => {
         console.log('error delete SQL INSERT', error)
         res.sendStatus(500);
-    });
+      });
+  } else {
+    res.sendStatus(403);
+  }
 });
 
+// PUT
 
 router.put('/:id', (req, res) => {
-  const newComment = req.body;
-  const queryText = `UPDATE comments
-  SET "comment" = $1 
-  WHERE id=$2;`; 
-  const queryValues = [
-    newComment.comment,
-    req.params.id,
-  ];
-  pool.query(queryText, queryValues)
-    .then(() => { res.sendStatus(200); })
-    .catch((err) => {
-      console.log('Error completing PUT comments query', err);
-      res.sendStatus(500);
-    });
+  if (req.isAuthenticated()) {
+    const newComment = req.body;
+    const queryText = `UPDATE comments
+                      SET "comment" = $1 
+                      WHERE id=$2;`;
+    const queryValues = [
+      newComment.comment,
+      req.params.id,
+    ];
+    pool.query(queryText, queryValues)
+      .then(() => { res.sendStatus(200); })
+      .catch((err) => {
+        console.log('Error completing PUT comments query', err);
+        res.sendStatus(500);
+      });
+  } else {
+    res.sendStatus(403);
+  }
 });
-
 
 module.exports = router;
